requirements.txt
# Requirements for the library project
Django==5.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
celery==5.3.1
django-cors-headers==4.2.0
pandas==2.1.0
matplotlib==3.8.0
seaborn==0.13.2
scikit-learn==1.3.0
beautifulsoup4==4.12.2
selenium==4.38.0
pytest==8.3.1
gunicorn==23.0.0
# AllAuth dependencies
django-allauth
python-dotenv
requests
pyjwt
cryptography
webdriver-manager
fake-useragent
tenacity
#groq
groq
# Embedding model dependencies
sentence-transformers
torch
numpy
